Updates calculation of volatile gas mass fractions
Usually we will not be able to account for all the volatile species predicted
by PCCL. To account for this we had previuosly just renormalized the fractions
(including tar) to equal 1.0 as they should in OF format. This is a problem
because it make the tar fraction unequal to that predicted in PCCL.

With the approach in this commit we normalize all of the other gas species
but leave the tar fraction unchanged. This ensures that the value for tar
fraction we generate in the primary volatile section will correspond to the
daf tar ultimate in tar breakdown equation.