Adds tar rate fields to STF
I want to have a little more detail on what the soot model is doing. This commit
adds rate fields, like those in the soot model.

I also added a field to track the number of tar explicit step iterations in a cell.
If the explicit integration works nicely there will be no over-consumption and there
will only be the specified number of sub steps (default in the model is currently
set to 3). This field enables us to see how well the model in general is behaving.