Changes parameters to SootTarModel
Previously we were passing volScalarFields for the rate fields. This is dumb
because it forces us to increment the ref counter within the submodel whenever
we want to get at the internal scalarField. Now we just pass a reference to
the internal scalarField as a parameter. This increments the ref coutner
but only once and it is an apporpriate ref counter increment.