Adds sootFlux field to the STF solver
The sooFlux is a vector field of unit [kg/(m^2 s). It is useful
when we are trying to determine a mass flow rate of soot to compare with
experimental results. This field, when integrated over a surface, yields
the mass flow rate. The vector needs to be dotted with surface normal of
course but in our case we just care about the axial component that is aligned
with the x axis, so we just integrate the vector and take the first component.