Replaces viscosity with thermal diffusivity for species
In the four main solvers, PCFSootTarFoam, NoGrowthSootTarFoam, SootTarFoam and
SootCoalFoam I replaced the species diffusivity, which was previously
set equal to the viscosity (i.e. a unity Schmidt number assumption), with the
thermal diffusivity (a unity Lewis number assumption). Obviously neither
one is really the mass diffusivity but Le=1 is a lot easier to
argue than Sc = 1.